Error Handling en methodes verder uitwerken

+++ src/domein/Fluvius.java
@@ -56,23 +56,25 @@ public class Fluvius
 	
 	public List<String> geefCategorien()
 	{
-        return getCategorien().stream().map(Categorie::toString).collect(Collectors.toList());
+		return getCategorien().stream().map(Categorie::toString).collect(Collectors.toList());
 	}
 	
 	public void setSdGoals()
 	{
-			sdGoals.clear();
-			sdGoals.addAll(sdGoalsRepo.findAll());
+		sdGoals.clear();
+		sdGoals.addAll(sdGoalsRepo.findAll());
 	}
-
+	
 	public ObservableList<SdGoal> getSdGoals()
 	{
+		setSdGoals();
+		System.out.println("Alle SdGoals ophalen");
 		return FXCollections.unmodifiableObservableList(sdGoals);
 	}
 	
 	public List<String> geefSdGoals()
 	{
-        return getSdGoals().stream().map(SdGoal::toString).collect(Collectors.toList());
+		return getSdGoals().stream().map(SdGoal::toString).collect(Collectors.toList());
 	}
 	
 	public void voegCategorieObserverToe(ListChangeListener<Categorie> listener)
@@ -101,12 +103,12 @@ public class Fluvius
 		}
 		catch(Exception e)
 		{
-			throw new IllegalArgumentException("");
+			throw new IllegalArgumentException("Er is een probleem opgetreden bij het toevoegen van een Categorie");
 		}
 		
 		setCategorien();
 	}
-
+	
 	//aangepast
 	public void voegCategorieToe(String naam, List<SdGoal> sdGoals, String icon)
 	{
@@ -116,36 +118,73 @@ public class Fluvius
 		voegCategorieToe(categorie);
 	}
 	
+	public void voegCategorieToe(String naam)
+	{
+		System.out.printf("Categorie %s aanmaken in java%n", naam);
+		Categorie categorie = new Categorie(naam);
+		System.out.printf("Categorie %s is aangemaakt in Java%n", categorie.toString());
+		voegCategorieToe(categorie);
+	}
+	
 	public void verwijderCategorie(Categorie categorie)
 	{
-		System.out.printf("Categorie %s verwijderen uit databank%n", categorie.toString());
-		GenericDaoJpa.startTransaction();
-		categorieRepo.delete(categorie);
-		GenericDaoJpa.commitTransaction();
+		try
+		{	
+			List<Categorie> categorien = categorieRepo.findAll();
+			if( categorien.size() == 1 && categorien.get(0).equals(categorie))
+			{
+				throw new IllegalArgumentException("Kan enigste categorie niet verwijderen");
+			}
+			
+			System.out.printf("Categorie %s verwijderen uit databank%n", categorie.toString());
+			GenericDaoJpa.startTransaction();
+			categorieRepo.delete(categorie);
+			GenericDaoJpa.commitTransaction();	
+		}
+		catch(IllegalArgumentException e)
+		{
+			throw e;
+		}
+		catch(Exception e)
+		{
+			throw new IllegalArgumentException("Er is een probleem opgetreden bij het verwijderen van een Categorie");
+		}
 		
 		setCategorien();
 	}
-
+	
 	public void verwijderCategorie(String naam)
 	{
 		System.out.printf("Categorie %s zoeken in databank%n", naam);
-		Categorie categorie = categorieRepo.getByNaam(naam);		
-		verwijderCategorie(categorie);		
-	}
-
-	public void wijzigCategorieNaam(Categorie categorie, String nieuweNaam)
-	{
-		categorie.setNaam(nieuweNaam);
-		updateCategorie(categorie);
+		Categorie categorie = categorieRepo.getByNaam(naam);
+		verwijderCategorie(categorie);
 	}
 	
 	public void updateCategorie(Categorie categorie)
 	{
-		GenericDaoJpa.startTransaction();
-		categorieRepo.update(categorie);
-		GenericDaoJpa.commitTransaction();
+		try
+		{
+			GenericDaoJpa.startTransaction();
+			categorieRepo.update(categorie);
+			GenericDaoJpa.commitTransaction();
+		}
+		catch(Exception e)
+		{
+			throw new IllegalArgumentException("Er is een probleem opgetreden bij een Categorie update");
+		}
 		
 		setCategorien();
+	}
+	
+	public void wijzigCategorieNaam(Categorie categorie, String nieuweNaam)
+	{
+		if(categorieRepo.getByNaam(nieuweNaam) != null)
+		{
+			throw new IllegalArgumentException("Er bestaad al een categorie met deze naam");
+		}
+		
+		categorie.setNaam(nieuweNaam);
+		updateCategorie(categorie);
 	}
 	
 	public void wijzigCategorieNaam(String oldName, String newName)
@@ -160,20 +199,20 @@ public class Fluvius
 //		getCategorien();
 //	}
 	
+	public void wijzigCategorieDoelstellingen(String naam, List<String> sdGoalsNamen)
+	{
+		Categorie categorie = categorieRepo.getByNaam(naam);
+		List<SdGoal> sdGoals = sdGoalsNamen.stream().map(sdgNaam -> sdGoalsRepo.getByNaam(sdgNaam))
+				.collect(Collectors.toList());
+		
+		wijzigCategorieDoelstellingen(categorie, sdGoals);
+	}
+	
 	public void wijzigCategorieDoelstellingen(Categorie categorie, List<SdGoal> sdGoals)
 	{
 		categorie.wijzigDoelstellingen(sdGoals);
 		updateCategorie(categorie);
 	}
-
-	public void wijzigCategorieDoelstellingen(String naam, List<String> sdGoalsNamen)
-	{
-		Categorie categorie = categorieRepo.getByNaam(naam);
-		List<SdGoal> sdGoals = sdGoalsNamen.stream().map(sdgNaam -> sdGoalsRepo.getByNaam(sdgNaam)).collect(Collectors.toList());
-		
-		wijzigCategorieDoelstellingen(categorie, sdGoals);
-	}
-
 	
 	public void setCategorieIcon(String categorieNaam, String icon)
 	{
@@ -187,8 +226,4 @@ public class Fluvius
 		Categorie categorie = categorieRepo.getByNaam(categorieNaam);
 		return categorie.getIcon();
 	}
-	
-	
 }
-
-
